Confirm and wait modal

+++ src/modules/daybook/views/EntryView.vue
@@ -37,6 +37,8 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
+
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
 export default {
@@ -76,6 +78,13 @@ export default {
     },
 
     async saveEntry() {
+      new Swal({
+        title: 'Espere por favor',
+        allowOutsideClick: false
+      })
+
+      Swal.showLoading()
+
       if (this.entry.id) {
 
         await this.updateEntry(this.entry)
@@ -84,11 +93,32 @@ export default {
 
         this.$router.push({ name: 'entry', params: { id } })
       }
+
+      Swal.fire('Guardado!', 'Entrada registrada con éxito', 'success')
     },
     async onDeleteEntry() {
-      await this.deleteEntry(this.entry.id)
+      const result = await Swal.fire({
+        title: 'Está seguro?',
+        text: 'Una vez borrado, no se puede recuperar',
+        showDenyButton: true,
+        confirmButtonText: 'Sí, estoy seguro!'
+      })
 
-      this.$router.push({ name: 'non-entry' })
+      if (result.isConfirmed) {
+        new Swal({
+          title: 'Espere, por favor',
+          allowOutsideClick: false
+        })
+
+        Swal.showLoading()
+        await this.deleteEntry(this.entry.id)
+
+        this.$router.push({ name: 'non-entry' })
+
+        Swal.fire('Eliminado con éxito', '', 'success')
+      }
+
+
     }
   },
   computed: {
